+++ Changes.md
@@ -6,6 +6,13 @@ Features
 
 - ContextQuery : Added node to access a context variable value directly without needing to use an expression.
 
+Improvements
+------------
+
+- Outputs :
+  - Added support for `layerName` string parameter, which can be used to customise the naming of channels in EXR outputs. Currently only supported for Arnold renders.
+  - Added support for `layerPerLightGroup` boolean parameter, which automatically splits the outputs into separate layers, one for each light group.
+
 API
 ---